Cleaning up the tests dramatically
A lot of the jank was due to me not realizing that I was ending my tests when the final planned assertion happened, and not when I was calling t.end.

Also dropped the bit of the testing that was trying to assert that my hashlocation.replace call would work in browsers the way I expected.